chore: Updates for Aug 2025
This PR contains a number of updates for a somewhat forgotten library.
I have shifted things to use edition 2024 (finally) and functions are
not quite so panicky anymore. Instead, functions return `VapidError`.

(There's probably a lot more that I can do to improve the code in here,
since it was one of the first things I wrote in rust.)